cleans up production build

+++ ui/scripts/build.sh
@@ -1,6 +1,15 @@
 #!/usr/bin/env bash
+export NODE_ENV=${NODE_ENV:-development}
+
+if [[ "$CI" != "" ]]; then 
+    export NODE_ENV=production
+fi
+
 npx snowpack build
-mkdir _build
-mv build/{*.js,*.html,*.map,assets} _build
-rm -rf build
-mv _build build
+
+if [[ "$CI" != "" ]]; then 
+    mkdir _build
+    mv build/{*.js,*.html,assets} _build
+    rm -rf build
+    mv _build build
+fi
